requirements2.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
catboost==0.24.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
cloudpickle==2.0.0
ConfigSpace==0.4.18
cycler==0.10.0
Cython==0.29.24
decorator==5.1.0
efficientnet-pytorch==0.6.3
fsspec==2021.10.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
graphviz==0.17
grpcio==1.41.0
hyperopt==0.2.2
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
liac-arff==2.5.0
lightgbm==3.1.0
Markdown==3.3.4
matplotlib==2.2.3
multidict==5.2.0
munch==2.5.0
networkx==2.2
numpy==1.21.2
oauthlib==3.1.1
openml==0.11.0
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
plotly==5.3.1
pretrainedmodels==0.7.4
progressbar33==2.4
protobuf==3.18.0
psutil==5.7.3
pthflops==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
pyrfr==0.8.2
python-dateutil==2.8.2
pytorch-lightning==1.4.9
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.7.1
segmentation-models-pytorch==0.2.0
simanneal==0.4.2
six==1.16.0
Surface-Distance-Based-Measures @ file:///export/scratch1/home/mmab/projects/SAGOMEA/surface-distance
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
timm==0.4.12
toposort==1.5
torch==1.8.1
torchmetrics==0.5.1
torchvision==0.9.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
xmltodict==0.12.0
yarl==1.6.3
